Reduce number of mallocs per send operation
Minimize number of memory allocations by storing the elements for the message
in the mailbox_element itself. Also, actors now simply transfer ownership of
the mailbox_element when forwarding a message. Calling last_dequeued() or
last_sender() while not in a message handler is now undefined behavior
(dereferences a nullptr).